pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stt-ai-agent"
version = "0.1.0"
description = "Local AI agent for speech-to-text, text processing, and knowledge base management"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "STT AI Agent", email = "naumannlinguae@gmail.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai-whisper",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "numpy>=1.20.0",
    "requests>=2.25.0",
    "python-dotenv>=0.19.0",
    "rich>=12.0.0",
    "typer>=0.7.0",
    "pydantic>=1.10.0",
    "chromadb>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
stt-agent = "stt_ai_agent.cli:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
